requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
crispy-tailwind==1.0.3
Django==5.0.4
django-appconf==1.0.6
django-compressor==4.4
django-crispy-forms==2.1
django-filter==24.2
# mysqlclient==2.2.4
pillow==10.3.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.5.0
tzdata==2024.1
